requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
idna==3.10
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
spotipy==2.25.0
sqlparse==0.5.3
urllib3==2.3.0
